Add disc classes for Far, Mall, and 13
