+++ src/main/java/goat/minecraft/minecraftnew/subsystems/music/discs/MusicDisc.java
@@ -1,0 +1,9 @@
+package goat.minecraft.minecraftnew.subsystems.music.discs;
+
+import org.bukkit.Material;
+import org.bukkit.entity.Player;
+
+public interface MusicDisc {
+    Material getDiscMaterial();
+    void onUse(Player player);
+}
